allows pinging midair now
NOT DONE!!!  when you try to ping midair, it will (if a variable is set)
ping 10 meters away from you in that direction.  However, the midair
ping is just a static final boolean, not configurable in some
configuration file (like it should be).  

attempted to make it be pick block instead of right click, but pick
block is apperantly clientside.  :(